AStyle reformating
Used astyle -R -A1 -f -U -xg -p -H -N -C -O *.cpp,*.h

Some refactoring

Some file renaming

Added initial doxyfile
