+++ src/cbSVPGProperties.cpp
@@ -16,7 +16,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/gpl-3.0>.
 */
 
-#include "cbsvpgproperties.h"
+#include "cbSVPGProperties.h"
 #include <wx/object.h>
 
 #include <cbcolourmanager.h>
@@ -35,13 +35,14 @@ IMPLEMENT_DYNAMIC_CLASS(svPGBaseProp, wxPGProperty)
 
 WX_PG_IMPLEMENT_VARIANT_DATA(svPGData)
 
-bool getDataFromVariant(const wxVariant& variant, svPGData& data )
+bool getDataFromVariant(const wxVariant& variant, svPGData& data)
 {
     uint64_t da = 0;
     wxString type = variant.GetType();
-    if(type == wxPG_VARIANT_TYPE_BOOL)
+
+    if (type == wxPG_VARIANT_TYPE_BOOL)
     {
-        if(variant.GetBool())
+        if (variant.GetBool())
             data.SetData(1);
         else
             data.SetData(0);
@@ -50,22 +51,23 @@ bool getDataFromVariant(const wxVariant& variant, svPGData& data )
     }
     else if (type == wxPG_VARIANT_TYPE_LONG)
     {
-         data.SetData(variant.GetLong());
-         return true;
+        data.SetData(variant.GetLong());
+        return true;
     }
     else if (type == wxT("svPGData"))
     {
         data = svPGDataRefFromVariant(variant);
         return true;
     }
+
     return false;
 };
 
 
 void svPGBaseProp::SetDataFromBinary(const wxString& str)
 {
-    svPGData& data= svPGDataRefFromVariant(m_value);
-    data.SetDataFromBinary(str ,GetMask(), GetBitOffset() );
+    svPGData& data = svPGDataRefFromVariant(m_value);
+    data.SetDataFromBinary(str, GetMask(), GetBitOffset());
     wxVariant var;
     var << data;
     SetValue(var, 0, 0);
@@ -84,7 +86,7 @@ void svPGBaseProp::SetData(uint64_t data)
 
 uint64_t svPGBaseProp::GetData()
 {
-    if(m_value.GetType() != wxT("svPGData"))
+    if (m_value.GetType() != wxT("svPGData"))
         return 0;
 
     svPGData& data = svPGDataRefFromVariant(m_value);
@@ -97,38 +99,45 @@ wxString svPGData::GetDataReadable(ValueRepresentation rep, uint64_t bitsize) co
     return GetDataReadable(m_data, bitsize, rep);
 }
 
-wxString svPGData::GetDataReadable(uint64_t data,uint64_t bitsize, ValueRepresentation rep) const
+wxString svPGData::GetDataReadable(uint64_t data, uint64_t bitsize, ValueRepresentation rep) const
 {
     wxString output;
-    switch(rep)
+
+    switch (rep)
     {
     case REP_BIN:
         output = wxT("0b") + printBits(bitsize, data);
         break;
+
     case REP_HEX:
     {
         wxString format;
-        format << wxT("0x%0") << bitsize/4 << wxT("llx");
-        output.Printf(format, data );
+        format << wxT("0x%0") << bitsize / 4 << wxT("llx");
+        output.Printf(format, data);
         break;
     }
+
     case REP_DEC:
-        output.Printf(wxT("%lli"), (int64_t) data );
+        output.Printf(wxT("%lli"), (int64_t) data);
         break;
+
     case REP_UDEC:
-        output.Printf(wxT("%llu"), data );
+        output.Printf(wxT("%llu"), data);
         break;
+
     case REP_FLOAT:
-        output.Printf(wxT("%f"), *(reinterpret_cast<const float*>(&data)) );
+        output.Printf(wxT("%f"), *(reinterpret_cast<const float*>(&data)));
         break;
+
     case REP_CHAR:
-        output.Printf(wxT("%c"), *(reinterpret_cast<const char*>(&data)) );
+        output.Printf(wxT("%c"), *(reinterpret_cast<const char*>(&data)));
         break;
+
     case REP_BYTE_ARRAY:
     {
-        for(size_t i = 0; i < bitsize / 8; ++i)
+        for (size_t i = 0; i < bitsize / 8; ++i)
         {
-            output += wxString::Format(wxT("0x%x "), *(reinterpret_cast<const char*>(&data + i)) );
+            output += wxString::Format(wxT("0x%x "), *(reinterpret_cast<const char*>(&data + i)));
         }
     }
     break;
@@ -147,7 +156,7 @@ void svPGData::SetDataFromBinary(const wxString& str, const uint64_t& mask, cons
     uint64_t data = 0;
 
     memcpy(&data, buff, size);
-    m_data = (data & mask)>>offset;
+    m_data = (data & mask) >> offset;
 }
 
 void svPGData::SetDataFromString(const wxString& str)
@@ -158,7 +167,7 @@ void svPGData::SetDataFromString(const wxString& str)
 WX_PG_IMPLEMENT_PROPERTY_CLASS(svPGPeripheryProp, svPGBaseProp,
                                svPGData, const svPGData&,
                                TextCtrl)
-svPGPeripheryProp::svPGPeripheryProp(SVDPeriphery &per) : svPGBaseProp( per )
+svPGPeripheryProp::svPGPeripheryProp(SVDPeriphery &per) : svPGBaseProp(per)
 {
     svPGData value(per.GetResetValue());
 
@@ -173,10 +182,12 @@ svPGPeripheryProp::svPGPeripheryProp(SVDPeriphery &per) : svPGBaseProp( per )
     desc << wxString::Format(wxT("Reset value: 0x%llx\n"), GetResetVal() & GetResetMask());
 
     auto itr = per.GetRegistersBegin();
-    for ( ; itr != per.GetRegistersEnd(); ++itr)
+
+    for (; itr != per.GetRegistersEnd(); ++itr)
     {
         SVDRegister* reg = dynamic_cast<SVDRegister*>((*itr).get());
-        if(reg == nullptr)
+
+        if (reg == nullptr)
             continue;
 
         size      += reg->GetSize() / 8;  // Size is in bytes
@@ -210,17 +221,20 @@ svPGPeripheryProp::~svPGPeripheryProp()
 void svPGPeripheryProp::Populate()
 {
     size_t child_count = GetChildCount();
-    for(size_t i = 0; i < child_count; ++i)
+
+    for (size_t i = 0; i < child_count; ++i)
     {
-        svPGRegisterProp *child = (svPGRegisterProp*) (Item(i));
-        if(child == nullptr)
+        svPGRegisterProp *child = (svPGRegisterProp*)(Item(i));
+
+        if (child == nullptr)
             continue;
+
         child->Populate();
     }
 }
 
 
-void svPGPeripheryProp::SetData(uint64_t data )
+void svPGPeripheryProp::SetData(uint64_t data)
 {
     // Not supported...
     LOG_UNEXPECTED_ERROR(wxT("Not supported in svPGPeripheryProp"));
@@ -256,24 +270,26 @@ void svPGPeripheryProp::RefreshChildren()
     size_t in_size = m_internalData.length();
     size_t start = 0;
 
-    for (unsigned int i = 0; i < GetChildCount() ; ++i )
+    for (unsigned int i = 0; i < GetChildCount() ; ++i)
     {
         svPGBaseProp* child = dynamic_cast<svPGBaseProp*>(Item(i));
-        if(child == nullptr)
+
+        if (child == nullptr)
         {
-            LOG_UNEXPECTED_ERROR( Item(i)->GetName() + _(" child of ") + GetName() + _(" can not be casted to svPGBaseProp*"));
+            LOG_UNEXPECTED_ERROR(Item(i)->GetName() + _(" child of ") + GetName() + _(" can not be casted to svPGBaseProp*"));
             continue;
         }
 
         size_t out_size = child->GetSize();
-        if(out_size > in_size-start)
+
+        if (out_size > in_size - start)
             out_size = in_size - start;
 
         char* target_buff = new char[out_size];
 
         memcpy(target_buff, buff.data() + start, out_size);
         start += out_size;
-        child->SetDataFromBinary(wxString::From8BitData(target_buff,out_size));
+        child->SetDataFromBinary(wxString::From8BitData(target_buff, out_size));
 
         delete[] target_buff;
     }
@@ -294,20 +310,22 @@ svPGRegisterProp::svPGRegisterProp(const SVDRegister &reg, const SVDPeriphery &p
     wxString desc = reg.GetDesc();
 
     auto itr = reg.m_fields.begin();
-    for ( ; itr != reg.m_fields.end(); ++itr)
+
+    for (; itr != reg.m_fields.end(); ++itr)
     {
         SVDField* field = dynamic_cast<SVDField*>((*itr).get());
-        if(field == nullptr)
+
+        if (field == nullptr)
             continue;
 
         wxPGProperty* prop;
 
-        if((field->m_bitRange.GetWidth() == 1) && (field->m_enumerated_value.GetValuesSize() == 0))
+        if ((field->m_bitRange.GetWidth() == 1) && (field->m_enumerated_value.GetValuesSize() == 0))
         {
             // Flag
             prop = new svPGBitProp(*field);
         }
-        else if(field->m_enumerated_value.GetValuesSize() != 0)
+        else if (field->m_enumerated_value.GetValuesSize() != 0)
         {
             // Enumeration
             prop = new svPGEnumFieldProp(*field);
@@ -318,7 +336,7 @@ svPGRegisterProp::svPGRegisterProp(const SVDRegister &reg, const SVDPeriphery &p
             prop = new svPGValueProp(*field);
         }
 
-        AddChild(prop);
+        AddPrivateChild(prop);
     }
 
     desc << wxString::Format(wxT("\nAddress: 0x%llx\n"), GetAddress());
@@ -342,6 +360,7 @@ uint64_t svPGRegisterProp::GetDataFromChildren()    const
 {
     uint64_t ret = 0;
     size_t count = GetChildCount();
+
     for (size_t i = 0; i < count; ++i)
     {
         const svPGData& data = svPGDataRefFromVariant(Item(i)->GetValue());
@@ -357,26 +376,33 @@ void svPGRegisterProp::Populate()
 {
 
     size_t child_count = GetChildCount();
-    for(size_t i = 0; i < child_count; ++i)
+
+    for (size_t i = 0; i < child_count; ++i)
     {
         wxPGProperty *child = Item(i);
-        if(child == nullptr)
+
+        if (child == nullptr)
             continue;
-        bool isKind = child->IsKindOf( CLASSINFO(svPGEnumFieldProp) );
-        if(isKind)
+
+        bool isKind = child->IsKindOf(CLASSINFO(svPGEnumFieldProp));
+
+        if (isKind)
         {
             svPGEnumFieldProp* en = (svPGEnumFieldProp*) child;
-            if(en == nullptr)
+
+            if (en == nullptr)
                 continue;
+
             en->Populate();
         }
     }
 }
 
-wxString svPGRegisterProp::ValueToString( wxVariant& value, int argFlags ) const
+wxString svPGRegisterProp::ValueToString(wxVariant& value, int argFlags) const
 {
     svPGData data;
-    if(getDataFromVariant(value, data) == false)
+
+    if (getDataFromVariant(value, data) == false)
         return wxEmptyString;
 
     return data.GetDataReadable(m_rep, GetBitSize());
@@ -395,11 +421,12 @@ void svPGRegisterProp::SetDataFromBinary(const wxString& str)
 }
 
 
-wxVariant svPGRegisterProp::ChildChanged( wxVariant& thisValue, int childIndex, wxVariant& childValue ) const
+wxVariant svPGRegisterProp::ChildChanged(wxVariant& thisValue, int childIndex, wxVariant& childValue) const
 {
 
     svPGData thisData;
-    if(getDataFromVariant(thisValue, thisData) == false)
+
+    if (getDataFromVariant(thisValue, thisData) == false)
     {
         LOG_UNEXPECTED_ERROR(wxT("thisValue can not be converted to svPGData"));
         return wxNullVariant;
@@ -407,7 +434,8 @@ wxVariant svPGRegisterProp::ChildChanged( wxVariant& thisValue, int childIndex, 
 
 
     svPGData childData;
-    if(getDataFromVariant(childValue, childData) == false)
+
+    if (getDataFromVariant(childValue, childData) == false)
     {
         LOG_UNEXPECTED_ERROR(wxT("thisValue can not be converted to childValue"));
         return wxNullVariant;
@@ -431,7 +459,7 @@ wxVariant svPGRegisterProp::ChildChanged( wxVariant& thisValue, int childIndex, 
     return newVariant;
 }
 
-void svPGRegisterProp::SetData( uint64_t data )
+void svPGRegisterProp::SetData(uint64_t data)
 {
     svPGData d;
     d.SetData(data);
@@ -442,27 +470,29 @@ void svPGRegisterProp::SetData( uint64_t data )
 
 void svPGRegisterProp::RefreshChildren()
 {
-    if(m_value.GetType() != wxT("svPGData"))
+    if (m_value.GetType() != wxT("svPGData"))
         return;
 
     svPGData& data = svPGDataRefFromVariant(m_value);
     int count = GetChildCount();
-    for(int i = 0; i < count; ++i)
+
+    for (int i = 0; i < count; ++i)
     {
         svPGBaseProp* child = dynamic_cast<svPGBaseProp*>(Item(i));
         svPGData child_data;
         const wxString& type = child->GetValue().GetType();
-        if (type == wxPG_VARIANT_TYPE_LONG )
+
+        if (type == wxPG_VARIANT_TYPE_LONG)
         {
-            child_data.SetData( child->GetValue().GetLong());
+            child_data.SetData(child->GetValue().GetLong());
         }
-        else if(type == wxT("svPGData") )
+        else if (type == wxT("svPGData"))
         {
             child_data = svPGDataRefFromVariant(child->GetValue());
         }
-        else if(type == wxPG_VARIANT_TYPE_BOOL)
+        else if (type == wxPG_VARIANT_TYPE_BOOL)
         {
-            child_data.SetData( child->GetValue().GetBool() == true  ? 1 : 0);
+            child_data.SetData(child->GetValue().GetBool() == true  ? 1 : 0);
         }
 
 
@@ -472,8 +502,8 @@ void svPGRegisterProp::RefreshChildren()
         uint64_t oldData = child_data.GetData();
         uint64_t newData = (data.GetData() & childMask) >>  childOffset;
 
-        if(oldData != newData)
-            child->ChangeFlag( wxPG_PROP_MODIFIED, true );
+        if (oldData != newData)
+            child->ChangeFlag(wxPG_PROP_MODIFIED, true);
 
         svPGData newPGData(newData);
         wxVariant var;
@@ -489,14 +519,14 @@ WX_PG_IMPLEMENT_PROPERTY_CLASS(svPGEnumFieldProp, svPGBaseProp,
                                svPGData, const svPGData&,
                                ComboBox)
 
-svPGEnumFieldProp::svPGEnumFieldProp(SVDField &field) : svPGBaseProp(field )
+svPGEnumFieldProp::svPGEnumFieldProp(SVDField &field) : svPGBaseProp(field)
 {
     svPGData value(field.GetResetValue());
 
     wxString desc =  field.GetDesc();
 
-    desc << wxString::Format(wxT("\nBit size: %lld\n"), GetBitSize() );
-    desc << wxString::Format(wxT("Bit offset: %lld\n"), GetBitOffset() );
+    desc << wxString::Format(wxT("\nBit size: %lld\n"), GetBitSize());
+    desc << wxString::Format(wxT("Bit offset: %lld\n"), GetBitOffset());
 
     SetDescription(desc);
     SetHelpString(desc);
@@ -506,6 +536,7 @@ svPGEnumFieldProp::svPGEnumFieldProp(SVDField &field) : svPGBaseProp(field )
     auto itr = field.m_enumerated_value.GetValuesBegin();
     size_t name_length = 0;
     size_t desc_length = 0;
+
     for (; itr != field.m_enumerated_value.GetValuesEnd(); ++itr)
     {
         name_length = std::max((*itr)->GetName().length(), name_length);
@@ -515,33 +546,37 @@ svPGEnumFieldProp::svPGEnumFieldProp(SVDField &field) : svPGBaseProp(field )
     name_length += 4;
 
     itr = field.m_enumerated_value.GetValuesBegin();
+
     for (; itr != field.m_enumerated_value.GetValuesEnd(); ++itr)
     {
         wxString name = (*itr)->GetName();
-        for(size_t i = name.length(); i < name_length; ++i)
+
+        for (size_t i = name.length(); i < name_length; ++i)
             name += _(" ");
 
-        m_elements.push_back( svPGEnumFieldElement((*itr)->GetName(),
-                              (*itr)->GetDescription(),
-                              name + wxT(": ") + (*itr)->GetDescription(),
-                              -1,
-                              (*itr)->GetValue()) );
+        m_elements.push_back(svPGEnumFieldElement((*itr)->GetName(),
+                             (*itr)->GetDescription(),
+                             name + wxT(": ") + (*itr)->GetDescription(),
+                             -1,
+                             (*itr)->GetValue()));
     }
 
     m_value << value;
 
 }
 
-wxString svPGEnumFieldProp::ValueToString( wxVariant& value, int argFlags ) const
+wxString svPGEnumFieldProp::ValueToString(wxVariant& value, int argFlags) const
 {
     wxString ret;
-    if(value.GetType() == wxT("svPGData") )
+
+    if (value.GetType() == wxT("svPGData"))
     {
         const svPGData& data = svPGDataRefFromVariant(value);
         auto itr = m_elements.begin();
-        for(; itr != m_elements.end(); ++itr)
+
+        for (; itr != m_elements.end(); ++itr)
         {
-            if(itr->value == data.GetData())
+            if (itr->value == data.GetData())
             {
                 ret = itr->text;
                 break;
@@ -552,23 +587,26 @@ wxString svPGEnumFieldProp::ValueToString( wxVariant& value, int argFlags ) cons
     else
     {
         long sel = value.GetLong();
-        if(sel >= 0 && sel < m_elements.size())
+
+        if (sel >= 0 && sel < m_elements.size())
             ret = m_elements[sel].text;
     }
+
     return ret;
 
 }
 
-bool svPGEnumFieldProp::StringToValue( wxVariant& variant, const wxString& text, int argFlags ) const
+bool svPGEnumFieldProp::StringToValue(wxVariant& variant, const wxString& text, int argFlags) const
 {
     svPGData& data = svPGDataRefFromVariant(variant);
     uint64_t old_data = data.GetData();
     uint64_t num = data.GetData();
 
     auto itr = m_elements.begin();
-    for(; itr != m_elements.end(); itr++)
+
+    for (; itr != m_elements.end(); itr++)
     {
-        if( itr->text == text)
+        if (itr->text == text)
         {
             num = itr->value;
             break;
@@ -577,9 +615,9 @@ bool svPGEnumFieldProp::StringToValue( wxVariant& variant, const wxString& text,
     }
 
 
-    data.SetData( num & (GetMask() >> GetBitOffset() ));  // Mask the data
+    data.SetData(num & (GetMask() >> GetBitOffset()));    // Mask the data
 
-    if(old_data == data.GetData())
+    if (old_data == data.GetData())
         return false;
     else
         return true;
@@ -588,6 +626,7 @@ bool svPGEnumFieldProp::StringToValue( wxVariant& variant, const wxString& text,
 void svPGEnumFieldProp::Populate()
 {
     auto itr = m_elements.begin();
+
     for (; itr != m_elements.end(); ++itr)
     {
         (*itr).index = AddChoice((*itr).text);
@@ -597,24 +636,27 @@ void svPGEnumFieldProp::Populate()
 int svPGEnumFieldProp::GetChoiceSelection() const
 {
     wxString type = m_value.GetType();
-    if(type == wxT("svPGData"))
+
+    if (type == wxT("svPGData"))
     {
         const svPGData& data = svPGDataRefFromVariant(m_value);
         uint64_t da = data.GetData();
         auto itr = m_elements.begin();
         size_t i = 0;
-        for(; itr != m_elements.end(); ++itr)
+
+        for (; itr != m_elements.end(); ++itr)
         {
-            if( itr->value == da)
+            if (itr->value == da)
                 return i;
+
             i++;
         }
     }
-    else if(type == wxPG_VARIANT_TYPE_LONG)
+    else if (type == wxPG_VARIANT_TYPE_LONG)
     {
         return m_value.GetLong();
     }
-    else if(type == wxPG_VARIANT_TYPE_BOOL)
+    else if (type == wxPG_VARIANT_TYPE_BOOL)
     {
         return m_value.GetBool() == true ? 1 : 0;
     }
@@ -623,10 +665,11 @@ int svPGEnumFieldProp::GetChoiceSelection() const
 }
 
 
-bool svPGEnumFieldProp::IntToValue( wxVariant &	variant, int number, int argFlags ) const
+bool svPGEnumFieldProp::IntToValue(wxVariant &	variant, int number, int argFlags) const
 {
     wxString type = variant.GetType();
-    if(type != wxT("svPGData"))
+
+    if (type != wxT("svPGData"))
     {
         svPGData data;
         variant << data;
@@ -634,14 +677,14 @@ bool svPGEnumFieldProp::IntToValue( wxVariant &	variant, int number, int argFlag
 
     svPGData& data = svPGDataRefFromVariant(variant);
 
-    if(number < 0 || number >= m_elements.size())
+    if (number < 0 || number >= m_elements.size())
         return false; //todo: ASSERT
 
     uint64_t oldData = data.GetData();
     uint64_t newData = m_elements[number].value;
     data.SetData(newData);
 
-    if(oldData != newData)
+    if (oldData != newData)
         return true;
     else
         return false;
@@ -671,14 +714,14 @@ svPGValueProp::svPGValueProp(const SVDField& field) : svPGBaseProp(field)
     m_value << value;
 }
 
-wxString svPGValueProp::ValueToString( wxVariant& value, int argFlags ) const
+wxString svPGValueProp::ValueToString(wxVariant& value, int argFlags) const
 {
     const svPGData& data = svPGDataRefFromVariant(value);
     wxString ret = data.GetDataReadable(m_rep, GetBitSize());
     return ret;
 }
 
-bool svPGValueProp::StringToValue( wxVariant& variant, const wxString& text, int argFlags ) const
+bool svPGValueProp::StringToValue(wxVariant& variant, const wxString& text, int argFlags) const
 {
     svPGData& data = svPGDataRefFromVariant(variant);
     uint64_t old_data = data.GetData();
@@ -686,19 +729,20 @@ bool svPGValueProp::StringToValue( wxVariant& variant, const wxString& text, int
 
     long long num;
     wxString number;
-    if(text.StartsWith(wxT("0x"),&number))
+
+    if (text.StartsWith(wxT("0x"), &number))
     {
-        number.ToLongLong(&num,16);
+        number.ToLongLong(&num, 16);
     }
     else
     {
-        text.ToLongLong(&num,10);
+        text.ToLongLong(&num, 10);
     }
 
 
-    data.SetData( num & (GetMask() >> GetBitOffset() ));  // Mask the data
+    data.SetData(num & (GetMask() >> GetBitOffset()));    // Mask the data
 
-    if(old_data == data.GetData())
+    if (old_data == data.GetData())
         return false;
     else
         return true;
@@ -710,14 +754,14 @@ WX_PG_IMPLEMENT_PROPERTY_CLASS(svPGBitProp, svPGBaseProp,
                                svPGData, const svPGData&,
                                CheckBox)
 
-svPGBitProp::svPGBitProp(const SVDField &field) : svPGBaseProp( field )
+svPGBitProp::svPGBitProp(const SVDField &field) : svPGBaseProp(field)
 {
     SetBitFlag(m_repCap, REP_HEX, REP_BIN, REP_DEC, REP_UDEC, REP_FLOAT, REP_CHAR);
     m_rep = REP_HEX;
 
     svPGData value(field.GetResetValue());
 
-    SetAttribute(wxT("UseCheckbox"),true);
+    SetAttribute(wxT("UseCheckbox"), true);
 
     wxString desc =  field.GetDesc();
 
@@ -734,24 +778,28 @@ svPGBitProp::svPGBitProp(const SVDField &field) : svPGBaseProp( field )
 int svPGBitProp::GetChoiceSelection() const
 {
     wxString type = m_value.GetType();
+
     if (type == wxT("svPGData"))
     {
         const svPGData& data = svPGDataRefFromVariant(m_value);
-        if(data.GetData())
+
+        if (data.GetData())
             return 1;
         else
             return 0;
 
-    } else if(type == wxPG_VARIANT_TYPE_BOOL)
+    }
+    else if (type == wxPG_VARIANT_TYPE_BOOL)
     {
         return m_value.GetBool();
     }
 }
 
-wxString svPGBitProp::ValueToString( wxVariant& value, int argFlags ) const
+wxString svPGBitProp::ValueToString(wxVariant& value, int argFlags) const
 {
     const svPGData& data = svPGDataRefFromVariant(value);
-    if(data.GetData())
+
+    if (data.GetData())
         return wxT("1");
     else
         return wxT("0");
@@ -772,53 +820,57 @@ void svPGBitProp::SetDataFromBinary(const wxString& str)
 
     memcpy(&data, buff, size);
 
-    svPGData& Data= svPGDataRefFromVariant(m_value);
+    svPGData& Data = svPGDataRefFromVariant(m_value);
 
-    Data.SetData((data & GetMask()) >> GetBitOffset() );
+    Data.SetData((data & GetMask()) >> GetBitOffset());
     wxVariant var;
     var << Data;
     SetValue(var, 0, 0);
 }
 
-bool svPGBitProp::StringToValue( wxVariant& variant, const wxString& text, int argFlags )
+bool svPGBitProp::StringToValue(wxVariant& variant, const wxString& text, int argFlags)
 {
     svPGData& data = svPGDataRefFromVariant(variant);
 
     uint64_t old_data = data.GetData();
     long long number;
 
-    if( (text.ToLongLong(&number) && number > 0) || text == wxT("1") || text == wxT("0x1") || text == wxT("true") || text == wxT("TRUE"))
+    if ((text.ToLongLong(&number) && number > 0) || text == wxT("1") || text == wxT("0x1") || text == wxT("true") || text == wxT("TRUE"))
         data.SetData(1);
     else if (text == wxT("0") || text == wxT("false") || text == wxT("FALSE"))
         data.SetData(0);
 
-    if(old_data == data.GetData())
+    if (old_data == data.GetData())
         return true;
     else
         return false;
 }
 
-bool svPGBitProp::IntToValue( wxVariant& variant, int number, int argFlags ) const
+bool svPGBitProp::IntToValue(wxVariant& variant, int number, int argFlags) const
 {
     wxString type = variant.GetType();
-    if(type != wxT("svPGData"))
+
+    if (type != wxT("svPGData"))
     {
         svPGData data;
         variant << data;
     }
 
     svPGData& data = svPGDataRefFromVariant(variant);
-    if(number == 0)
+
+    if (number == 0)
         data.SetData(0);
     else
         data.SetData(1);
+
     return true;
 }
 
 void svPGBitProp::SetChoiceSelection(int newValue)
 {
     svPGData& data = svPGDataRefFromVariant(m_value);
-    if(newValue == 0)
+
+    if (newValue == 0)
         data.SetData(0);
     else
         data.SetData(1);
@@ -831,12 +883,13 @@ wxString printBits(size_t const size, uint64_t data)
     ret.clear();
     int i, j;
 
-    j = size -1;
+    j = size - 1;
 
-    for ( ; j>=0; j--)
+    for (; j >= 0; j--)
     {
         byte = (data >> j) & 1;
-        if(byte)
+
+        if (byte)
             ret << wxT("1");
         else
             ret << wxT("0");
